Commit for version 0.1, which also includes some really old code
This doesn't have EventTime stuff in it and the elog data is too
large to be worth including in this commit.  The next version will
have a fairly different way of storing and generating the data,
and will focus on z-position data (rather than wait times).

Also included is the stuff in the v0 folders, which I believe is
the code from before I decided to do an object-oriented approach
(so that stuff is pretty ancient).

This version of the code should work pretty well; I've spent a
lot of time debugging it.  The documentation is a little outdated
but I expect no one to really use this version, and the next
version will likely be fairly different, so it would not be worth
it for me to update the documentation right now.  You're better
off using a newer version.