First commit for the new setup. Lots and lots of changes.
The code has only been lightly tested, so it probably still has
some bugs.

Currently, nothing is parallelized, the documentation is not
updated, and the cmb velocity calculations still return results in
Cartesian coordinates (rather than giving the speed and relative
angle of the trap).

Random numbers are still generated by matlab's rand function using
its Mersenne twister algorithm, but the seed is set (instead of
just keeping the Matlab default of 0) and is indirectly based off
of the current time.

z-positions are picked (with replacement) from the tracer output.