Performed minor updates on Analysis.m
Fixed up printing during Analysis.generate_Charman_histograms() to
make it similar to the other status output.  Also set it to only
add the default signal_groups if not signal_group file was found
on the hard drive.  Furthmore took out one pointless statement
that measured the size of an array for debugging, which
I accidentally left behind after deleting the print statement.