Added a new time generator based on shift schedule
Currently this algorithm generates roughly twice as many events
as it should. This could very likely be due to the value for
n_Hbar_lambda, which is currently set to one event per 1.306
hours.

Furthermore, right now all events are assumed to be quip right.
This should change in the future.