Fixed minor bug with parallel use of the random number generators.
Copies of the random number generators are passed to workers in
an spmd structure.  This means that the state of the generators in
the analysis parent doesn't change when the generators are called
in the spmd loop.  Now the analysis parent random generators are
synchronized with the generators in the workers after leaving the
spmd loop in Signal_Group.run().