Fixed bug in Analysis.generate_Charman_tables() x-axis
The issue that Joel had noted earlier in which the x-axis seemed
to give values that were an order of magnitude too large.  This
happened because the old version used the same bin centers for
each signal group when plotting, but through an error did not use
the same bins when calculating the bin heights.

There still seems to be a bug somewhere in the data processing.
When analyzed, the pure_sine signal groups give an amplitude of
half of what is expected.