Modified CharmanI*.m to return the amplitude of the signal
Previously, the two functions returned the A_1 parameter as it was
given in Andy's paper.  However, for the given frequency omega,
there is also a -omega component frequency with complex amplitude
A_-1=conj(A_1).  Adding these contributions together gives a
cosine wave with amplitude 2*abs(A_1) and phase arg(A_1), i.e.
2*abs(A_1)*cos(omega*t+arg(A_1)) (and yes that's a plus, not a
minus in front of the phase).  This is the same as taking the real
part of a complex signal with 2*A_1*exp(i*omega*t).  Therefore the
A_1 returned by the Charman functions is now twice the A_1 given
by the formulas in the paper he sent around.