Added TracerDataSet class and EllipToMat.m
TracerDataSet is designed to hold the data from simulations for
one charge and one quip direction.

EllipToMat.m is designed to take a collected output file from the
server and turn it into a sorted .mat file that the
PositionGenerator class will be able to use once it is updated.