Fixed bug with Analysis.save_signal_group_list() and added Analysis.set_all_n_sets()
There was an issue in which function handles in signal_group
instances contained a pointer to the analysis parent due to the
way in which Matlab saves workspaces into function handles.  This
could cause issues when saving the analysis.save_signal_group_list.
These issues occured when calling that function after calling
analysis.run().  That issues is now fixed by creating the function
handles in a namespace that doesn't contain a reference to any
analysis instances (this occurs in a static method of the analysis
class).  Also moved analysis.add_signal_group() to the hidden
methods section so that users don't create function handles in the
interpretter which will likely have the analysis instance in its
namespace, causing these issues.  That function should probably
only be called internally with its function handle argument being
created in a static instance with a clean namespace.

Analysis.set_all_n_sets() iterates over analysis.signal_group_list
and sets n_sets for each of them to the same given value.