Fixed normalization on Charman histograms and issue with adding signals
Charman histograms are now normalized so that the area underneath
them is constant (previously they were normalized so that the bins
summed to one, but that doesn't account for differing bin width
for different signals).

Fixed issue where adding sine signals with different amplitudes
was not allowed because both of their func2str values were the
same due to the fact that 'amplitude' was printed instead of the
actual charge value e.g. '1.000e-8'.