Added CMB_Speed_Sinusoidal_Charge_Comparison.m and Test_Charman.m
The first plots normalized curves showing the variation of CMB
speed of the Earth and the fractional charge in our model as
functions of time.  This shows how the CMB speed varies in a
very nearly sinusoidal manner.

The second is a work in progress that creates a sine wave and
fits it using CharmanII or CharmanIV then plots the results.
Right now there is some phase difference between the two plots
that I believe is due to the fact that the Charman algorithms
take phase to be zero for a cosine wave that starts at datenum(0)
instead of for a cosine wave in phase with the CMB speed.