Included some scripts and edits for plots for the paper
Included script and plots that demonstrate the CDF interpolation
procedure.  Also included the plot and corresponding script that
displays the Earth CMB Speed over the course of a few years with
the Hbar events marked.

Furthermore edited Position_Generator.m to allow the user to
speicfy the random value used by the interpolator, which is
useful for debugging and plotting.