Fixed small bug in Position_Generator
After adding the feature to select the random number used in the
interpolation (useful for plotting and debugging), the ability to
choose a random number on its own was broken.  This was due to the
fact that the way in which it determined whether or not the
optional rand_val argument was given was incorrect, so it always
assumed that argument was provided.  This is now fixed.